Add TextCommandsContext#getEffectivePrefixes
