+++ src/main/kotlin/io/github/freya022/botcommands/internal/commands/text/TextCommandsListener.kt
@@ -41,7 +41,6 @@ private val spacePattern = Regex("\\s+")
 @ConditionalService(TextCommandsListener.ActivationCondition::class)
 internal class TextCommandsListener internal constructor(
     private val context: BContext,
-    private val textPrefixSupplier: TextPrefixSupplier?,
     private val defaultMessagesFactory: DefaultMessagesFactory,
     private val textCommandsContext: TextCommandsContextImpl,
     private val localizableTextCommandFactory: LocalizableTextCommandFactory,
@@ -173,26 +172,9 @@ internal class TextCommandsListener internal constructor(
     }
 
     private fun getMsgNoPrefix(msg: String, guild: Guild): String? {
-        return getPrefixes(guild)
+        return textCommandsContext.getEffectivePrefixes(guild)
             .find { prefix -> msg.startsWith(prefix) }
             ?.let { prefix -> msg.substring(prefix.length).trimStart() }
-    }
-
-    private fun getPrefixes(guild: Guild): List<String> {
-        if (textPrefixSupplier != null) {
-            return textPrefixSupplier.getPrefixes(guild)
-        }
-
-        context.settingsProvider?.let { settingsProvider ->
-            val prefixes = settingsProvider.getPrefixes(guild)
-            if (!prefixes.isNullOrEmpty()) return prefixes
-        }
-
-        val prefixes = context.textConfig.prefixes
-        return when (context.textConfig.usePingAsPrefix) {
-            false -> prefixes
-            true -> prefixes + guild.selfMember.asMention
-        }
     }
 
     private suspend fun canRun(event: MessageReceivedEvent, commandInfo: TextCommandInfo): Boolean {
